Move LocalStack fixtures into the main conftest
