--- tests/localstack/__init__.py
@@ -1,3 +1,0 @@
-from __future__ import annotations
-
-__all__: list[str] = []
